requirements.txt
Owlready2~=0.35
rdflib~=6.0.2
nltk==3.6.3
numpy~=1.21.4
matplotlib~=3.4.3
PyYAML~=6.0
spacy==3.1.4
allennlp==2.8.0
allennlp-models==2.8.0